Link to Canonical Path version of Posts
We want our links to "look pretty" and be consistent

Here we add a trait to convert a PathBuf to a canonical path,
and then use that in the blog index page.

This deals with stripping the `md` extension of the full `index.md`
file, whichever is relevant.

Fixes #14

It does create a NEW problem where sometimes images aren't found
due to the directory structure