Do some Refactoring
I started by looking to see if we could remove the
`article_path: Option<String>` and replace it with
a `String` that we knew always existed.

That led me to the usages in `til_get` and others
where we were using `None`, but in theory could have
the relative url of that page, since I render markdown
for more pages than just plain blog posts.

I found that I had already made the `LinkTo` trait
that helped make Absolute URLs, for my different types
of pages!

Once I found that crate I leaned in to using it. I
used that got the blog post plus all the other places
we render markdown.

I also realized that in my `absolute_link` method I
was manually making a url with strings, and I didn't
like that. So I replaced it with using the `Url` crate.

But I also didn't like that I had to parse the `config.base_url`
`String` into a `Url` every time I wanted to use it

So I changed the `AppConfig` to have a `Url`
instead of a `String`. This refactor was easier than
I thought and helped clean up another link related
code path.