ImgProxy CDN (#66)
* Get imgproxy setup, but since I have local paths to images I need to get the full url here which requires piping context down farther

* fixed the the image url ref for imgproxy

* fixing some lint issues

* fixing lint issues in cmd.rs

* added the MarkdownRenderContext struct and add changes to the imports related to SyntaxHighlightingContext

* used PathBuf for more robust path handling

* fixed rebasing errors

* Get imgproxy setup, but since I have local paths to images I need to get the full url here which requires piping context down farther

* fixed the the image url ref for imgproxy

* fixing some lint issues

* fixing lint issues in cmd.rs

* added the MarkdownRenderContext struct and add changes to the imports related to SyntaxHighlightingContext

* used PathBuf for more robust path handling

* fixed rebasing errors

* Do some Refactoring

I started by looking to see if we could remove the
`article_path: Option<String>` and replace it with
a `String` that we knew always existed.

That led me to the usages in `til_get` and others
where we were using `None`, but in theory could have
the relative url of that page, since I render markdown
for more pages than just plain blog posts.

I found that I had already made the `LinkTo` trait
that helped make Absolute URLs, for my different types
of pages!

Once I found that crate I leaned in to using it. I
used that got the blog post plus all the other places
we render markdown.

I also realized that in my `absolute_link` method I
was manually making a url with strings, and I didn't
like that. So I replaced it with using the `Url` crate.

But I also didn't like that I had to parse the `config.base_url`
`String` into a `Url` every time I wanted to use it

So I changed the `AppConfig` to have a `Url`
instead of a `String`. This refactor was easier than
I thought and helped clean up another link related
code path.

* I didn't like this struct, its not worth it I feel like to make this struct just to use this one method from the crate, its easy enough to use the config mehtod directly

* clippy

* Delete the tailwind binary

* passing img url to imgproxy url

* removed /bin/bash from the repo

* Remove binary tailwindcss file and add it to .gitignore

* deleted tailwindcss bin file

* added .DS to gitignore and added the suggested refactor on the imgproxy url path

* Use the env var if set for the imgproxy base url, and cleanup some

* Use a relative url unless imgproxy is setup

* More commented out lines

* Save markdown var so we don't have to call method multiple times

* Just set the base url correctly the first time. Does mean we have to guess to format of the fly url but I think thats ok

* Use one img tag with a srcset

---------

Co-authored-by: Seif-Mamdouh <seifmamdouh7878@gmail.com>